+++ prequ/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '0.300.0.post0.dev0'
+__version__ = '0.400.0'